requirements.txt
torch
torchaudio
torchvision
easyocr
requests
opencv-python==4.6.0.66
tqdm
matplotlib
pillow
tensorboardX
numpy